@@ -1,22 +1,27 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <Version>0.6</Version>
+    <TargetFramework>net5.0</TargetFramework>
+    <Version>0.6.9</Version>
     <PackAsTool>true</PackAsTool>
     <ToolCommandName>SQLCover</ToolCommandName>
     <PackageOutputPath>./nupkg</PackageOutputPath>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="CommandLineParser" Version="2.8.0" />
-    <PackageReference Include="Microsoft.Data.SqlClient" Version="1.1.1" />
-    <PackageReference Include="Microsoft.SqlServer.DACFx" Version="150.4573.2" />
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="3.0.0" />
+    <PackageReference Include="Microsoft.SqlServer.DACFx" Version="150.5164.1" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\SQLCoverLib\SQLCoverLib.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="nupkg\" />
   </ItemGroup>
 
 </Project>
