@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
 using System.Linq;
 using System.Threading;
+using Microsoft.Data.SqlClient;
 using SQLCover.Gateway;
 using SQLCover.Source;
 using SQLCover.Trace;
@@ -135,7 +135,7 @@ namespace SQLCover
             {
                 _database.Execute(command, timeOut); //todo read messages or rowcounts or something
             }
-            catch (System.Data.SqlClient.SqlException e)
+            catch (SqlException e)
             {
                 if (e.Number == -2)
                 {
